Changed to depend on optional arguments because it was too concrete

+++ susvm.py
@@ -317,6 +317,8 @@ def _start_app_intercept(app):
         if app.poll() == 0: break
 
 def start(args):
+    move, is_full = args_parse(args)
+
     vers = glob.glob(rf'{APP_VERSION_PATH}\ver.*')
     if not vers:
         print('The version you want to use is not selected.')
@@ -326,10 +328,11 @@ def start(args):
     ver = vers[0]
 
     app = unstable_app_open(cmd = APP_NAME, cwd = ver, wait = 2)
-    type_keys('win+shift+right')
+    if move: type_keys(f'win+shift+{move}')
 
-    helper_app = unstable_app_open(HELPER_APP_PATH, wait = 2)
-    helper_app.kill()
+    if is_full:
+        helper_app = unstable_app_open(HELPER_APP_PATH, wait = 2)
+        helper_app.kill()
 
     print(f'{version(ver)} is running now')
 
@@ -364,6 +367,8 @@ parser_use.add_argument('ver', help='Specify the version to use')
 parser_use.set_defaults(handler=use)
 
 parser_start = subparsers.add_parser(TYPE_START, help='Start the app based on the version used')
+parser_start.add_argument('-m', '--move', help='Moves the application window between monitors')
+parser_start.add_argument('-f', '--full', action='store_true', help='Make the application window virtual full screen')
 parser_start.set_defaults(handler=start)
 
 args = parser.parse_args()
